+++ sklearn/metrics/__init__.py
@@ -42,7 +42,7 @@ from .cluster import normalized_mutual_info_score
 from .cluster import fowlkes_mallows_score
 from .cluster import silhouette_samples
 from .cluster import silhouette_score
-from .cluster import calinski_harabaz_score
+from .cluster import calinski_harabasz_score
 from .cluster import v_measure_score
 
 from .pairwise import euclidean_distances